Better plugin interface and javadoc.
